refactor(markdown): extract shared Marked renderer + content splitter
Create src/lib/markdown.ts with createMarkdownRenderer() and
splitMarkdownContent(). Both Prose.svelte and benny/+page.server.ts
now import from the shared module, eliminating ~60 lines of duplication.
(findings 2.1)
